Add query title selection feature and read notes in the subdirectories
If there is a match in the title, it automatically selects the title. However, this does not considers auto-filling. Without this auto-filling, we cannot discern between a note selection and a note creating. If we implement auto-filling later, then the non-initial partial title matching filter should be turned off.

Other changes that is not related to this CL.
- Sort notes by modified time.
- Delete debug messages.